# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: my_customTools API
  description: |
    媒体转码工具集后端 API 文档。

    包含视频转码、图片处理、动图转换、文档转换等功能。

    ## 认证方式
    使用 JWT Bearer Token 认证，在请求头中添加：
    ```
    Authorization: Bearer <token>
    ```

    ## 响应格式
    所有接口统一返回格式：
    ```json
    {
      "code": 0,
      "msg": "ok",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: 开发环境
  - url: http://production-server/api
    description: 生产环境
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证
  schemas:
...
    ApiResponse:
      type: object
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          description: 业务状态码，0 表示成功
          example: 0
        msg:
          type: string
          description: 响应消息
          example: "ok"
        data:
          type: object
          description: 响应数据
    PaginatedData:
      type: object
      properties:
        list:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            totalPages:
              type: integer

    # ========== 用户相关 ==========
...
    TaskStatus:
      type: string
      enum: [waiting, uploading, processing, completed, failed, cancelled]
    TaskType:
      type: string
      enum: [video, img, anim, document]
    BaseTask:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/TaskType'
        userId:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        inputPath:
          type: string
        outputPath:
          type: string
        inputFormat:
          type: string
        outputFormat:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        config:
          type: object
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        errorMsg:
          type: string
    CreateTaskRequest:
      type: object
      required:
        - type
        - fileName
        - inputPath
        - outputFormat
      properties:
        type:
          $ref: '#/components/schemas/TaskType'
        fileName:
          type: string
        fileSize:
          type: integer
        inputPath:
          type: string
        inputFormat:
          type: string
        outputFormat:
          type: string
        config:
          oneOf:
            - $ref: '#/components/schemas/VideoTranscodeConfig'
            - $ref: '#/components/schemas/ImgTranscodeConfig'
            - $ref: '#/components/schemas/AnimTranscodeConfig'

    # ========== 转码配置 ==========
    VideoTranscodeConfig:
      type: object
      properties:
        videoCodec:
          type: string
          enum: [h264, h265, vp9, av1, copy]
        resolution:
          type: string
          example: "1920x1080"
        bitrate:
          type: string
          example: "4M"
        crf:
          type: integer
          minimum: 0
          maximum: 51
        fps:
          type: integer
        audioCodec:
          type: string
          enum: [aac, mp3, opus, copy]
        audioBitrate:
          type: string
          example: "128k"
        startTime:
          type: number
        endTime:
          type: number
        rotate:
          type: integer
          enum: [90, -90, 180]
        hwAccel:
          type: string
          enum: [nvidia, vaapi, videotoolbox]
    ImgTranscodeConfig:
      type: object
      properties:
        outputFormat:
          type: string
          enum: [jpg, jpeg, png, webp, avif, bmp, tiff, ico, heic]
        quality:
          type: integer
          minimum: 1
          maximum: 100
        lossless:
          type: boolean
        resizeMode:
          type: string
          enum: [none, width, height, scale, fixed, crop]
        width:
          type: integer
        height:
          type: integer
        keepAspect:
          type: boolean
        colorSpace:
          type: string
          enum: [original, srgb, rgb, cmyk, grayscale]
        stripMeta:
          type: boolean
        stripICC:
          type: boolean
        rotate:
          type: integer
          enum: [0, 90, 180, 270]
        flipH:
          type: boolean
        flipV:
          type: boolean
        background:
          type: string
        compression:
          $ref: '#/components/schemas/ImageCompressionConfig'
    AnimTranscodeConfig:
      type: object
      properties:
        outputFormat:
          type: string
          enum: [gif, webp, apng]
        fps:
          type: integer
        width:
          type: integer
        height:
          type: integer
        colors:
          type: integer
        quality:
          type: integer
        loop:
          type: integer
        dither:
          type: string
          enum: [bayer, floyd, none]
        optimize:
          type: boolean
        delay:
          type: integer
        startTime:
          type: number
        duration:
          type: number
        imagePaths:
          type: array
          items:
            type: string
    ImageCompressionConfig:
      type: object
      properties:
        mode:
          type: string
          enum: [quality, size, percent]
        targetSize:
          type: integer
          description: 目标大小（字节）
        targetPercent:
          type: integer
          minimum: 1
          maximum: 100
        minQuality:
          type: integer
          minimum: 1
          maximum: 100

    # ========== 文档转换 ==========
...
paths:
...
  /tasks:
    post:
      tags: [Tasks]
      summary: 创建任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误

    get:
      tags: [Tasks]
      summary: 查询任务列表
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TaskType'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedData'
...
  /tasks/compose-gif:
    post:
      tags: [Tasks]
      summary: 创建 GIF 合成任务
      description: 将图片序列合成为 GIF/WebP/APNG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - imagePaths
                - outputFormat
              properties:
                imagePaths:
                  type: array
                  items:
                    type: string
                  minItems: 1
                outputFormat:
                  type: string
                  enum: [gif, webp, apng]
                config:
                  $ref: '#/components/schemas/AnimTranscodeConfig'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: 获取任务详情
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '404':
          description: 任务不存在

    delete:
      tags: [Tasks]
      summary: 删除任务
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
        '400':
          description: 删除失败
  /tasks/{taskId}/status:
    put:
      tags: [Tasks]
      summary: 更新任务状态
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/TaskStatus'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误
  /tasks/{taskId}/cancel:
    post:
      tags: [Tasks]
      summary: 取消任务
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 取消失败
  /tasks/{taskId}/retry:
    post:
      tags: [Tasks]
      summary: 重试任务
      description: 重新排队失败的任务
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 重试成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 重试失败
  /tasks/{taskId}/submit:
    post:
      tags: [Tasks]
      summary: 提交任务到队列
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 提交成功
        '400':
          description: 提交失败

  # ==================== Queue ====================
...
  /queue/pause/{type}:
    post:
      tags: [Queue]
      summary: 暂停队列
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [video, img, anim]
      responses:
        '200':
          description: 暂停成功
        '400':
          description: 参数错误
  /queue/resume/{type}:
    post:
      tags: [Queue]
      summary: 恢复队列
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [video, img, anim]
      responses:
        '200':
          description: 恢复成功
        '400':
          description: 参数错误
  /queue/clear/{type}:
    delete:
      tags: [Queue]
      summary: 清空队列
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [video, img, anim]
      responses:
        '200':
          description: 清空成功
        '400':
          description: 参数错误

  # ==================== Download ====================
...
  /clean/tasks:
    post:
      tags: [Clean]
      summary: 清理旧任务
      description: 需要管理员权限
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                maxAgeDays:
                  type: integer
                  default: 7
                  description: 清理多少天前的任务
      responses:
        '200':
          description: 清理成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tasksRemoved:
                            type: integer
        '403':
          description: 权限不足
...
  /export/tasks:
    get:
      tags: [Export]
      summary: 导出任务报告
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TaskType'
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 导出文件
          content:
            application/json:
              schema:
                type: string
            text/csv:
              schema:
                type: string
        '400':
          description: 参数错误
...